fix schema and refactor the code
