+++ api/src/main/java/org/apache/iceberg/Scan.java
@@ -133,4 +133,19 @@ interface Scan<T extends Scan<T>> {
    * @return an Iterable of tasks for this scan
    */
   CloseableIterable<CombinedScanTask> planTasks();
+
+  /**
+   * Returns the target split size for this scan.
+   */
+  long targetSplitSize();
+
+  /**
+   * Returns the split lookback for this scan.
+   */
+  int splitLookback();
+
+  /**
+   * Returns the split open file cost for this scan.
+   */
+  long splitOpenFileCost();
 }
