+++ api/src/main/java/org/apache/iceberg/TableScan.java
@@ -112,19 +112,4 @@ public interface TableScan extends Scan<TableScan> {
    * @return true if case sensitive, false otherwise.
    */
   boolean isCaseSensitive();
-
-  /**
-   * Returns the target split size for this scan.
-   */
-  long targetSplitSize();
-
-  /**
-   * Returns the split lookback for this scan.
-   */
-  int splitLookback();
-
-  /**
-   * Returns the split open file cost for this scan.
-   */
-  long splitOpenFileCost();
 }
